Align integer type

+++ nasdaq-taq/src/main/java/org/jvirtanen/nasdaq/taq/TAQSink.java
@@ -34,7 +34,7 @@ class TAQSink implements Flushable, MarketListener {
         timestampLow  = 0;
     }
 
-    public void timestamp(int high, long low) {
+    public void timestamp(long high, long low) {
         timestampHigh = high;
         timestampLow  = low;
     }
